requirements.txt
alembic==1.2.1
argon2-cffi==19.2.0
astroid==2.2.5
blinker==1.4
boto3==1.19.1
Click==7.0
Flask==1.1.1
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
flask-talisman==1.0.0
Flask-WTF==0.14.2
gunicorn==22.0.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.2
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.1
mccabe==0.6.1
Pillow>=6.2.2
psycopg2-binary==2.8.4
PyJWT==2.0.1
pylint==2.3.1
pytest==5.2.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-slugify==4.0.0
pytz==2019.3
schedule==0.6.0
six==1.12.0
SQLAlchemy==1.3.7
SQLAlchemy-Utils==0.36.1
typed-ast==1.4.0
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
